+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
+-   Fixed an issue when exporting numbers to excel. Instead of the type number, the value exported was of string type.
+
 -   Fixed an issue where pagination buttons wouldn't show up when there's no items in the Datagrid 2 pagination even when button visibility was set to "always".
 
 ### Added